Major Updates
- initial integration of sensor backend api
- added password reset
- integration of login, verify, widget blades to the SPA
- fixes & improvements

+++ app/Http/Controllers/FloorsController.php
@@ -16,9 +16,9 @@ class FloorsController extends Controller
     public function get(Request $request) {
         // by building
         if ($request->bid) {
-            $bid = Hashids::decode($request->bid)[0];
+            // $bid = Hashids::decode($request->bid)[0];
             
-            return response(Floor::where('building_id', $bid)->get());
+            return response(Floor::where('building_id', $request->bid)->get());
         }
 
         // by id
@@ -32,41 +32,39 @@ class FloorsController extends Controller
         return response(Floor::all());
     }
 
-    public function getData(Request $request, $id) {
-        $fid = Hashids::decode($id)[0];
+    // public function getData(Request $request, $id) {
+    //     $fid = Hashids::decode($id)[0];
 
-        // sensors only
-        if ($request->so) {
-            return response([
-                'sensors' => SensorMap::with('area')->where('floor_id', $fid)->get()
-            ]);
-        }
+    //     // sensors only
+    //     if ($request->so) {
+    //         return response([
+    //             'sensors' => SensorMap::with('area')->where('floor_id', $fid)->get()
+    //         ]);
+    //     }
 
-        return response([
-            'areas' => AreaMap::where('floor_id', $fid)->get(),
-            'sensors' => SensorMap::with('area')->where('floor_id', $fid)->get()
-        ]);
-    }
+    //     return response([
+    //         'areas' => AreaMap::where('floor_id', $fid)->get(),
+    //         'sensors' => SensorMap::with('area')->where('floor_id', $fid)->get()
+    //     ]);
+    // }
 
     public function create(Request $request) {
-        if ($request->building == '' || !$request->has('building')) {
-            return response(['r' => false, 'm' => 'Building is required']);
-        } else if ($request->floor_no == '' || !$request->has('floor_no')) {
-            return response(['r' => false, 'm' => 'Floor number is required']);
+        if ($request->building_id == '' || !$request->has('building_id')) {
+            return response(['r' => false, 'm' => 'Building reference is required']);
+        } else if ($request->ref_id == '' || !$request->has('ref_id')) {
+            return response(['r' => false, 'm' => 'Floor reference is required']);
         } else {
-            $bid = Hashids::decode($request->building)[0];
+            //$bid = Hashids::decode($request->building)[0];
             if (Floor::where([
-                ['building_id', $bid],
-                ['floor_no', $request->floor_no]
+                ['building_id', $request->building_id],
+                ['ref_id', $request->ref_id]
             ])->count() > 0) {
-                return response(['r' => false, 'm' => 'Floor '.$request->floor_no.' already exist']);
+                return response(['r' => false, 'm' => 'Floor already exist']);
             } else {
                 $floor = new Floor;
 
-                $floor->building_id = $bid;
-                $floor->floor_no = $request->floor_no;
-                $floor->size_metre = $request->size_metre;
-                $floor->size_feet = $request->size_feet;
+                $floor->building_id = $request->building_id;
+                $floor->ref_id = $request->ref_id;
                 $floor->occupancy_limit = $request->occupancy_limit;
                 $floor->save();
 
@@ -76,34 +74,20 @@ class FloorsController extends Controller
     }
 
     public function update(Request $request, $id) {
-        if ($request->floor_no == '' || !$request->has('floor_no')) {
-            return response(['r' => false, 'm' => 'Floor number is required']);
-        } else {
-            $fid = Hashids::decode($id)[0];
-            $floor = Floor::find($fid);
+        $fid = Hashids::decode($id)[0];
+        $floor = Floor::find($fid);
 
-            if (Floor::where([
-                    ['building_id', $floor->building_id],
-                    ['floor_no', $request->floor_no],
-                    ['id', '<>', $fid]
-                ])->count() > 0) {
-                return response(['r' => false, 'm' => 'Floor '.$request->floor_no.' already exist']);
-            } else {
-                $floor->floor_no = $request->floor_no;
-                $floor->size_metre = $request->size_metre;
-                $floor->size_feet = $request->size_feet;
-                $floor->occupancy_limit = $request->occupancy_limit;
-                $floor->save();
+        $floor->occupancy_limit = $request->occupancy_limit;
+        $floor->save();
 
-                return response(['r' => true, 'm' => 'Floor updated']);
-            }
-        }
+        return response(['r' => true, 'm' => 'Floor updated']);
     }
 
     /* Floor plan upload */
     public function setFloorPlan(Request $request) {
         if ($request->hasFile('floor_plan')) {
-            $fid = Hashids::decode($request->id)[0];
+            // $fid = Hashids::decode($request->id)[0];
+            $fid = $request->id;
 
             try {
                 $fpFolder = 'floors';
@@ -130,15 +114,23 @@ class FloorsController extends Controller
                 $thumb->save($thumbPath);
 
                 // update floor
-                $floor = Floor::find($fid);
+                // $floor = Floor::find($fid);
+                $floor = Floor::where('ref_id', $fid)->first();
 
-                // clean up old floor plan images
-                if ($floor->floor_plan) {
-                    Storage::delete($fpFolder.'/'.$floor->floor_plan);
-                    Storage::delete($fpThumbFolder.'/'.$floor->floor_plan);
+                if ($floor) {
+                    // clean up old floor plan images
+                    if ($floor->floor_plan) {
+                        Storage::delete($fpFolder.'/'.$floor->floor_plan);
+                        Storage::delete($fpThumbFolder.'/'.$floor->floor_plan);
+                    }
+
+                    $floor->floor_plan = $filename;
+                } else {
+                    $floor = new Floor;
+                    $floor->building_id = $request->bid;
+                    $floor->ref_id = $fid;
+                    $floor->occupancy_limit = $request->occupancy_limit;
                 }
-
-                $floor->floor_plan = $filename;
                 $floor->save();
 
                 return response(['r' => true, 'm' => 'Floor plan uploaded', 'floor_plan' => $filename]);
@@ -151,8 +143,9 @@ class FloorsController extends Controller
     }
 
     public function delete($id) {
-        $fid = Hashids::decode($id)[0];
-        $floor = Floor::find($fid);
+        // $fid = Hashids::decode($id)[0];
+        // $floor = Floor::find($fid);
+        $floor = Floor::where('ref_id', $id)->first();
 
         if ($floor) {
             $floor->delete();
